Extract storage classes to separate file
